Create Ubuntu-16.10 Unity Launcher Customiser
This script customises the launcher in Unity automatically for you. You simply add a list of your preferred apps and it will add them to the launcher. It will remove useless icons like the Amazon launcher, but preserve the useful ones such as the Desktop icon. 

It can be improved by adding settings to a configuration file instead of directly in the script. Additionally, it only supports applications, it can be further improved to support other kinds of launcher icons, as well as custom .desktop files.